- Changed timestamp datatype to long - Changed logic to avoid reading newer data elements than instances - Reflected table changes in text and application

+++ CosmosToPostgreSQL/CosmosToPostgreSQL/CosmosDataElement.cs
@@ -13,7 +13,7 @@ namespace CosmosToPostgreSQL
     {
         [JsonProperty(PropertyName = "_ts")]
         [JsonPropertyName("_ts")]
-        public int Ts { get; set; }
+        public long Ts { get; set; }
 
         [JsonProperty(PropertyName = "fileScanResult")]
         [JsonPropertyName("FileScanResult")]
